Continue development

+++ start-deps.sh
@@ -1,7 +1,15 @@
 #!/bin/sh
+if test -z $1;then
+  echo 'start-deps.sh PREFIX'
+  exit 1
+fi
+
 docker kill $1-db
 docker kill $1-mq
 docker kill $1-adminmongo
+
+docker volume create $1-db
+docker volume create $1-mq
 
 docker run --rm -d -v "${1}-db":/data/mongo -p 27017:27017 --name "${1}-db" mongo:4
